Change buflen to uint16_t type

+++ max30102_for_stm32_hal.c
@@ -5,7 +5,7 @@ void max30102_init(max30102_t *obj, I2C_HandleTypeDef *hi2c)
     obj->ui2c = hi2c;
 }
 
-void max30102_write(max30102_t *obj, uint8_t reg, uint8_t *buf, size_t buflen)
+void max30102_write(max30102_t *obj, uint8_t reg, uint8_t *buf, uint16_t buflen)
 {
     uint8_t *payload = (uint8_t *)malloc((buflen + 1) * sizeof(uint8_t));
     *payload = reg;
@@ -15,7 +15,7 @@ void max30102_write(max30102_t *obj, uint8_t reg, uint8_t *buf, size_t buflen)
     free(payload);
 }
 
-void max30102_read(max30102_t *obj, uint8_t reg, uint8_t *buf, size_t buflen)
+void max30102_read(max30102_t *obj, uint8_t reg, uint8_t *buf, uint16_t buflen)
 {
     uint8_t reg_addr = reg;
     HAL_I2C_Master_Transmit(obj->ui2c, MAX30102_I2C_ADDR << 1, &reg_addr, 1, MAX30102_I2C_TIMEOUT);
